requirements.txt
python-dotenv
PyYAML
websockets
pandas
numpy
fastapi
uvicorn

aiohttp
pydantic
pytest
openai
faiss-cpu
